transfer commit

+++ Assets/Scripts/Audio/AudioFilter.cs
@@ -4,18 +4,18 @@ public class AudioFilter : MonoBehaviour
 {
     float lowestFrequency;
     float highestFrequency;
-    double[] fftSamples;
-    public AudioFilter(float _lowestFrequency, float _highestFrequency, double[] _fftSamples)
+    float[] fftSamples;
+    public AudioFilter(float _lowestFrequency, float _highestFrequency, float[] _fftSamples)
     {
         lowestFrequency = _lowestFrequency;
         highestFrequency = _highestFrequency;
         fftSamples = _fftSamples;
     }
-    public double[] HighAndLowPassFilter()
+    public float[] HighAndLowPassFilter()
     {
-        double[] filteredFFT = new double[fftSamples.Length];
-        double[] highpass = HighPassFilter(lowestFrequency, fftSamples);
-        double[] lowpass = LowPassFilter(highestFrequency, fftSamples);
+        float[] filteredFFT = new float[fftSamples.Length];
+        float[] highpass = HighPassFilter(lowestFrequency, fftSamples);
+        float[] lowpass = LowPassFilter(highestFrequency, fftSamples);
         int cutofBinLowPass = (int)(highestFrequency * fftSamples.Length * 2 / NoteManager.Instance.DefaultSamplerate);
         int cutofBinHighPass = (int)(lowestFrequency * fftSamples.Length * 2 / NoteManager.Instance.DefaultSamplerate);
 
@@ -36,7 +36,7 @@ public class AudioFilter : MonoBehaviour
         }
         return filteredFFT;
     }
-    double[] AllPassFilter(float _cutofFrequency, double[] _fftSamples)
+    float[] AllPassFilter(float _cutofFrequency, float[] _fftSamples)
     {
         float fb = _cutofFrequency;
         float fs = NoteManager.Instance.DefaultSamplerate;
@@ -44,18 +44,18 @@ public class AudioFilter : MonoBehaviour
         float b = Mathf.Tan((Mathf.PI * fb) / fs);
         float a = b - 1 / b + 1;
 
-        double[] output = new double[_fftSamples.Length];
+        float[] output = new float[_fftSamples.Length];
         for (int i = 0; i < _fftSamples.Length; i++)
         {
             output[i] = (a + Mathf.Pow(i, -1)) / (1 + (a * Mathf.Pow(i, -1)));
         }
         return output;
     }
-    public double[] HighPassFilter(float _cutofFrequency, double[] _fftSamples)
+    public float[] HighPassFilter(float _cutofFrequency, float[] _fftSamples)
     {
         int length = _fftSamples.Length;
-        double[] filterdFFT = new double[length];
-        double[] allpass = AllPassFilter(_cutofFrequency, _fftSamples);
+        float[] filterdFFT = new float[length];
+        float[] allpass = AllPassFilter(_cutofFrequency, _fftSamples);
 
         for (int i = 0; i < length; i++)
         {
@@ -67,13 +67,14 @@ public class AudioFilter : MonoBehaviour
         //{
         //    filterdFFT[i] = 0;
         //}
+        
 
         return filterdFFT;
 
     }
-    public double[] LowPassFilter(float _cutofFrequency, double[] _fftSamples)
+    public float[] LowPassFilter(float _cutofFrequency, float[] _fftSamples)
     {
-        double[] filterdFFT = _fftSamples;
+        float[] filterdFFT = _fftSamples;
         int cutofBin = (int)(_cutofFrequency * _fftSamples.Length * 2 / NoteManager.Instance.DefaultSamplerate);
         for (int i = cutofBin; i < _fftSamples.Length; i++)
         {
